wip: progress on automation, some cleanup

+++ packages/lambda/scripts/build-binaries.sh
@@ -1,48 +1,58 @@
 #!/bin/sh
 # shellcheck shell=dash
 
-# Requires jq and Docker
+#
+# Builds specified or all headless browsers (chromium, firefox) version defined in package.json
+#
+# Requires Docker, jq, and zip
+#
+# Usage: ./build-binaries.sh [chromium|firefox]
+#
 
 set -e
 
-PACKAGE_DIRECTORY="$(dirname "$0")/.."
+cd "$(dirname "$0")/.."
 
+PACKAGE_DIRECTORY=$(pwd)
 
 build() {
   BUILD_NAME=$1
-  DOCKER_IMAGE=serverless-chrome-$BUILD_NAME
-  VERSION=$(jq -r .config."$BUILD_NAME"Version package.json)
-  BUILD_PATH=build/$BUILD_NAME
-  TARBALL_PATH=$BUILD_PATH/../headless-$BUILD_NAME-$VERSION-amazonlinux-2017-03.tar.gz
-  
-  export VERSION
-  
-  echo "t: $TARBALL_PATH"
 
   cd "$PACKAGE_DIRECTORY/builds/$BUILD_NAME"
-
-  exit
-
-  mkdir -p "$BUILD_PATH"
-
-  # Build
-  docker build -t "$DOCKER_IMAGE" .
-
-  # Extract headless_shell from docker image
-  docker run -dt --name "$DOCKER_IMAGE" "$DOCKER_IMAGE"
-  docker cp "$DOCKER_IMAGE":/chrome/Chromium/src/out/Headless/headless_shell "$BUILD_PATH"
-  docker stop "$DOCKER_IMAGE"
-
-  # Package
-  tar --no-xattrs --hard-dereference -cznshf "$TARBALL_PATH" "$BUILD_PATH"
-
-
-  # Cleanup
-  rm -Rf "$BUILD_PATH"
-
   
+  DOCKER_IMAGE=serverless-chrome-$BUILD_NAME
+  # VERSION=$(jq -r .config."$BUILD_NAME"Version "$PACKAGE_DIRECTORY"/package.json)
+  VERSION=$(./latest.sh)
+  BUILD_PATH=build/$BUILD_NAME
+  ZIPFILE_PATH=$BUILD_PATH/../headless-$BUILD_NAME-$VERSION-amazonlinux-2017-03.zip
+    
+  if [ ! -f "$ZIPFILE_PATH" ]; then
+    export VERSION
+
+    echo "Building $BUILD_NAME version $VERSION"
+    
+    mkdir -p "$BUILD_PATH"
+
+    # Build
+    docker build -t "$DOCKER_IMAGE" --build-arg VERSION="$VERSION" .
+
+    # Extract binary from docker image
+    docker run -dt --name "$DOCKER_IMAGE" "$DOCKER_IMAGE"
+    docker cp "$DOCKER_IMAGE":/build/headless-"$BUILD_NAME" "$BUILD_PATH"
+    docker stop "$DOCKER_IMAGE"
+
+    # Package
+    zip -9 "$ZIPFILE_PATH" "$BUILD_PATH"
+
+    # Cleanup
+    rm -Rf "$BUILD_PATH"
+    
+  else
+    echo "$BUILD_NAME version $VERSION was previously built. Skipping build."
+  fi
 }
 
+# main script
 
 if [ ! -z "$1" ]; then
   build "$1"
@@ -53,6 +63,3 @@ else
     build "${DOCKER_FILE%%/*}"
   done
 fi
-
-
-exit
